temporarily disable temporary file generation option on win32 platform.
The problem is that win32 api of GetTempFileNameW has a very high chance of confliction which means we cannot safely use it.

However #include<immintrin.h> takes 0.5 seconds of compilation time which is absurd on GCC.

GCC provides its own intrinsics for x86_64, we need to use that instead of using intel's intrinsics. Or the compilation time bloats ridiculously.