posix_file and c_file forgot to add the logic for constructing from win32_family file
W-series APIs of NT kernel was a huge mistake for Windows XP. Windows 95, 98 and ME did this better. But Windows XP totally screwed this up.
I know a lot of people would argue windows xp was the best windows os ever. However, it made a lot of historical mistakes. Like the updating systems, activation systems, removing the real mode DOS and NTVDM was a trash.
W series apis totally screwed the compatibility of programs between Win9X and Windows XP. And the freaking encoding on Windows XP, 7 and 10 are horribly bad. Microsoft finally admitted they fucked this up in 2019 and windows 10 finally tried some fixes, but it is already too late.

That is FUCKING windows XP's fault.