overhual implementation of intrinsics.h
The problem is that GCC, clang and msvc completely disagree with this for everything. Even so gcc does not agree with itself. One example is that -march=native for MSDOS when we run on x86_64 host, it will create tons of compilation errors.

The correct way is to guard every intrinsics with __has_builtin to prevent all sorts of problems.